HardGameFlaskApp/requirements.txt
Flask>=2.2.2
werkzeug==2.2.2
gunicorn==20.1.0
scikit-learn==0.24.2
numpy==1.19.5
pandas==1.2.4
pymongo==3.11.4
coremltools==4.1
google-cloud-storage==1.42.3
joblib==1.0.1
protobuf==3.20.*
dnspython==2.0.0